+++ robot-description-builder/src/chained/chained_jointbuilder.rs
@@ -89,13 +89,13 @@ mod tests {
 					let tree = Link::builder("Leg_[L1]_l1")
 						.add_visual(
 							Visual::builder(geom_leg_l1.clone())
-								.tranformed(Transform::new_translation(0., 1.5, 0.))
+								.transformed(Transform::new_translation(0., 1.5, 0.))
 								.named("Leg_[L1]_l1_vis_1")
 								.materialized(material_l1.clone()),
 						)
 						.add_collider(
 							Collision::builder(geom_leg_l1.clone())
-								.tranformed(Transform::new_translation(0., 1.5, 0.))
+								.transformed(Transform::new_translation(0., 1.5, 0.))
 								.named("Leg_[L1]_l1_col_1"),
 						)
 						.build_tree();
@@ -107,7 +107,7 @@ mod tests {
 							Link::builder("Leg_[L1]_l2")
 								.add_visual(
 									Visual::builder(geom_leg_l2.clone())
-										.tranformed(Transform::new(
+										.transformed(Transform::new(
 											(0., 5., 0.),
 											(FRAC_PI_2, 0., 0.),
 										))
@@ -116,7 +116,7 @@ mod tests {
 								)
 								.add_collider(
 									Collision::builder(geom_leg_l2.clone())
-										.tranformed(Transform::new(
+										.transformed(Transform::new(
 											(0., 5., 0.),
 											(FRAC_PI_2, 0., 0.),
 										))