+++ robot-description-builder/src/link/builder/collision_builder.rs
@@ -53,7 +53,7 @@ impl CollisionBuilder {
 	/// Specify a `transform` for this `CollisionBuilder`.
 	///
 	/// The default is a no transformation (The frame of the `Collision` will be the same as the frame of the parent `Link`).
-	pub fn tranformed(mut self, transform: Transform) -> Self {
+	pub fn transformed(mut self, transform: Transform) -> Self {
 		self.transform = Some(transform);
 		self
 	}